Refactored file structure. Created empty Form component.
