+++ src/App.js
@@ -1,13 +1,19 @@
 import React, { useState } from "react";
-import data from "./data";
+import data from "./data/data";
 
-import "./App.css";
+import "./styles/App.css";
+
+import Form from "./components/Form";
 
 function App() {
   const [teamList, updateTeamList] = useState(data);
-  console.log(data);
+  console.log("Team List: ", teamList);
 
-  return <div className="App" />;
+  return (
+    <div className="App">
+      <Form />
+    </div>
+  );
 }
 
 export default App;
